add_col_spanners {papaja}	R Documentation

Add table headings to group columns

Description

Takes a named list of containing column numbers to group with a heading This function is not exported.

Usage

add_col_spanners(table_lines, col_spanners, n_cols)

Arguments

table_lines
Character. Vector of characters containing one line of a LaTeX table each.

col_spanners
List. A named list containing the indices of the first and last columns to group, where the names are the headings.

n_cols
Numeric. Number of columns of the table.

See Also

apa_table

Examples

NULL

[Package papaja version 0.1.0.9997 Index]