add_effect_sizes {papaja}	R Documentation

Effect sizes for Analysis of Variance

Description

Calculates effect-size measures for Analysis of Variance output objects.

Usage

add_effect_sizes(x, es = "ges", observed = NULL, mse = TRUE, intercept = FALSE)

Arguments

x
An object of class apa_variance_table.

es
Character. A vector naming all to-be-computed effect-size measures. Currently, partial eta-quared ("pes"), generalized eta-squared ("ges"), and eta-squared ("es") are supported.

observed
Character. A vector naming all factors that are observed (i.e., not manipulated).

mse
Logical. Should means-squared errors be computed?

intercept
Logical. Should the sum of squares of the intercept (i.e., the deviation of the grand mean from 0) be included in the calculation of eta-squared?

[Package papaja version 0.1.0.9997 Index]