add_row_names {papaja}	R Documentation

Add row names as first column

Description

Adds row names as the first column of the table and sets row.name attribute to NULL. This function is not exported.

Usage

add_row_names(x, added_stub_head, force = FALSE)

Arguments

x
data.frame or matrix.

added_stub_head
Character. Used as stub head (name of first column).

force
Logical. Should row names be added even if they are the numbers 1:nrow(x)?

See Also

apa_table

Examples

NULL

[Package papaja version 0.1.0.9997 Index]