apa6_pdf {papaja}	R Documentation

APA article (6th edition)

Description

Template for creating an article according to APA guidelines (6th edition) in PDF format.

Usage

apa6_pdf(
  fig_caption = TRUE,
  number_sections = FALSE,
  toc = FALSE,
  keep_tex = TRUE,
  md_extensions = NULL,
  includes = NULL,
  ...
)

apa6_docx(fig_caption = TRUE, md_extensions = NULL, ...)

apa6_word(fig_caption = TRUE, md_extensions = NULL, ...)

Arguments

fig_caption
TRUE to render figures with captions

number_sections
TRUE to number section headings

toc
TRUE to include a table of contents in the output

keep_tex
Keep the intermediate tex file used in the conversion to PDF

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

includes
Named list of additional content to include within the document (typically created using the includes function).

...
Further arguments to pass to pdf_document or word_document.

Details

When creating PDF documents the YAML option classoption is passed to the class options of the LaTeX apa6 document class. In this case, additional options are available. Refer to the apa6 document class documentation to find out about class options such as paper size or draft watermarks.

When creating PDF documents the output device for figures defaults to c("pdf", "png"), so that each figure is saved in all four formats at a resolution of 300 dpi.

Functions

apa6_docx: Format to create .docx-files. class parameter is ignored. This function should be considered experimental.

apa6_word: Format to create .docx-files. Alias of apa6_docx.

See Also

html_document2, pdf_document, word_document

Examples

NULL

[Package papaja version 0.1.0.9997 Index]