apa_factorial_plot_single {papaja}	R Documentation

Plots for factorial designs that conform to APA guidelines, two-factors internal function

Description

Internal function that is called (possibly multiple times) by apa_factorial_plot.

Usage

apa_factorial_plot_single(
  aggregated,
  y.values,
  id,
  dv,
  factors,
  intercept = NULL,
  ...
)

Arguments

aggregated
A data.frame, the aggregated data.

y.values
A data.frame containing the measures of central tendency and of dispersion per cell of the design.

id
Character. Variable name that identifies subjects.

dv
Character. The name of the dependent variable.

factors
Character. A vector of up to four variable names that is used to stratify the data.

intercept
Numeric. See details in apa_factorial_plot

[Package papaja version 0.1.0.9997 Index]