apa_prepare_doc {papaja}	R Documentation

Prepare APA document (deprecated)

Description

Prepares objects to be used in the rendering process and creates title page and abstract for MS Word documents. This function is no longer needed and thus has been deprecated.

Usage

apa_prepare_doc()

Details

The function creates and locks a non-exported object apa_lang that is used by other papaja-functions. apa_lang is a list containing localizations for document elements such as abstract and title. The selected language is defined by the lang-parameter in the documents yaml-header. Currently, English (default) and German ("german") are supported.

If the output document is MS Word (output: apa6_word) the function additionally creates a title page and adds the abstract. You should, therefore, always call apa_prepare_doc at the very beginning of the R Markdown document.

See Also

apa6_word

[Package papaja version 0.1.0.9997 Index]