apa_print {papaja}	R Documentation

Format statistics (APA 6th edition)

Description

A generic function that takes objects from various statistical methods to create formatted character strings to report the results in accordance with APA manuscript guidelines. The function invokes particular methods, which depend on the class of the first argument.

Usage

apa_print(x, ...)

Arguments

x
Output object. See details.

...
Additional arguments passed to other methods. See details.

Details

Currently the following output objects are supported:

htest

lm and summary.lm

aov, aovlist, summary.aov, and summary.aovlist

anova and Anova.mlm

Value

The form of the value returned by apa_print depends on the class of x. See the documentation of the particular methods for details (e.g., ?apa_print.htest).

See Also

Other apa_print: apa_print.BFBayesFactor(), apa_print.aov(), apa_print.emmGrid(), apa_print.glht(), apa_print.glm(), apa_print.htest(), apa_print.list()

Examples

# List methods for apa_print()
methods("apa_print")

[Package papaja version 0.1.0.9997 Index]