apa_print.aov {papaja}	R Documentation

Format statistics from ANOVA (APA 6th edition)

Description

These methods take objects from various R functions that calculate ANOVA to create formatted character strings to report the results in accordance with APA manuscript guidelines. For anova-objects from model comparisons see apa_print.list.

Usage

## S3 method for class 'aov'
apa_print(x, ...)

## S3 method for class 'summary.aov'
apa_print(x, ...)

## S3 method for class 'aovlist'
apa_print(x, ...)

## S3 method for class 'summary.aovlist'
apa_print(x, ...)

## S3 method for class 'Anova.mlm'
apa_print(
  x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE,
  ...
)

## S3 method for class 'summary.Anova.mlm'
apa_print(
  x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE,
  ...
)

## S3 method for class 'afex_aov'
apa_print(
  x,
  correction = getOption("papaja.sphericity_correction"),
  intercept = FALSE,
  ...
)

## S3 method for class 'anova'
apa_print(x, ...)

## S3 method for class 'manova'
apa_print(x, test = "Pillai", ...)

Arguments

x
Output object. See details.

...
Arguments passed on to print_anova

es

Character. The effect-size measure to be calculated; can be either ges for generalized eta-squared, pes for partial eta-squared or es for eta-squared. Note that eta-squared is calculated correctly if and only if the design is balanced.

mse

Logical. Indicates if mean squared errors should be included in output. Default is TRUE.

observed

Character. The names of the factors that are observed, (i.e., not manipulated). Necessary for calculation of generalized eta-squared; otherwise ignored.

in_paren

Logical. Indicates if the formated string will be reported inside parentheses. See details.

correction
Character. In the case of repeated-measures ANOVA, the type of sphericity correction to be used (GG for Greenhouse-Geisser or HF for Huyn-Feldt methods or none). Default is GG.

intercept
Logical. Indicates if intercept test should be included in output.

test
For MANOVA, the multivariate test statistic to be reported, see summary.manova.

Details

The factor names are sanitized to facilitate their use as list names (see Value section). Parentheses are omitted and other non-word characters are replaced by _.

If in_paren is TRUE parentheses in the formatted string, such as those surrounding degrees of freedom, are replaced with brackets.

Value

apa_print.aov and related functions return a named list containing the following components according to the input:

statistic

A named list of character strings giving the test statistic, parameters, and p value for each factor.

estimate

A named list of character strings giving the effect size estimates for each factor.

full_result

A named list of character strings comprised of estimate and statistic for each factor.

table

A data.frame containing the complete ANOVA table, which can be passed to apa_table.

References

Bakeman, R. (2005). Recommended effect size statistics for repeated measures designs. Behavior Research Methods , 37 (3), 379–384. doi:10.3758/BF03192707

See Also

aov, Anova, apa_print.list

Other apa_print: apa_print.BFBayesFactor(), apa_print.emmGrid(), apa_print.glht(), apa_print.glm(), apa_print.htest(), apa_print.list(), apa_print()

Examples

   ## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   apa_print(npk_aov)

[Package papaja version 0.1.0.9997 Index]