apa_print.emmGrid {papaja}	R Documentation

Format statistics (APA 6th edition)

Description

Takes various lsmeans and emmeans objects to create formatted character strings to report the results in accordance with APA manuscript guidelines. These methods are not properly tested and should be considered experimental.

Usage

## S3 method for class 'emmGrid'
apa_print(x, ...)

## S3 method for class 'summary_emm'
apa_print(x, contrast_names = NULL, est_name = NULL, in_paren = FALSE, ...)

## S3 method for class 'lsmobj'
apa_print(x, ...)

## S3 method for class 'summary.ref.grid'
apa_print(x, ...)

Arguments

x
Object adaptation, such as χ^2 instead of "x-squared") is used for the test statistic, otherwise the

...
Arguments passed on to printnum

contrast_names
Character. An optional vector of names to identify calculated contrasts.

est_name
Character. If NULL (default) the name is guessed from the function call of the model object passed to lsmeans/emmeans.

in_paren
Logical. Indicates if the formated string will be reported inside parentheses.

Details

ADJUSTED CONFIDENCE INTERVALS

If in_paren is TRUE parentheses in the formated string, such as those surrounding degrees of freedom, are replaced with brackets.

Value

apa_print() returns a list containing the following components according to the input:

statistic

A character string giving the test statistic, parameters (e.g., degrees of freedom), and p value.

estimate

A character string giving the descriptive estimates and confidence intervals if possible

full_result

A joint character string comprised of est and stat.

table

A data.frame containing the complete contrast table, which can be passed to apa_table.

See Also

Other apa_print: apa_print.BFBayesFactor(), apa_print.aov(), apa_print.glht(), apa_print.glm(), apa_print.htest(), apa_print.list(), apa_print()

Examples

   NULL

[Package papaja version 0.1.0.9997 Index]