apa_print.glht {papaja}	R Documentation

Format statistics (APA 6th edition)

Description

Takes various lsmeans and emmeans objects methods to create formatted character strings to report the results in accordance with APA manuscript guidelines. These methods are not properly tested and should be considered experimental.

Usage

## S3 method for class 'glht'
apa_print(x, test = multcomp::adjusted(), ...)

## S3 method for class 'summary.glht'
apa_print(x, ci = 0.95, in_paren = FALSE, ...)

Arguments

x
Object

test
Function. Computes p-values (adjusted for multiple comparisons).

...
Further arguments to pass to printnum to format the estimate.

ci
Numeric. If NULL (default) the function tries to obtain confidence intervals from x. Other confidence intervals can be supplied as a vector of length 2 (lower and upper boundary, respectively) with attribute conf.level, e.g., when calculating bootstrapped confidence intervals.

in_paren
Logical. Indicates if the formated string will be reported inside parentheses.

Details

ADJUSTED CONFIDENCE INTERVALS

If in_paren is TRUE parentheses in the formated string, such as those surrounding degrees of freedom, are replaced with brackets.

Value

apa_print() returns a list containing the following components according to the input:

statistic

A character string giving the test statistic, parameters (e.g., degrees of freedom), and p value.

estimate

A character string giving the descriptive estimates and confidence intervals if possible

full_result

A joint character string comprised of est and stat.

table

A data.frame containing the complete contrast table, which can be passed to apa_table.

See Also

Other apa_print: apa_print.BFBayesFactor(), apa_print.aov(), apa_print.emmGrid(), apa_print.glm(), apa_print.htest(), apa_print.list(), apa_print()

Examples

   NULL

[Package papaja version 0.1.0.9997 Index]