apa_print_container {papaja}	R Documentation

Create empty container for results

Description

Creates the default empty container for the results of apa_print. This function is not exported.

Usage

apa_print_container()

Value

A named list containing the following components according to the input:

estimate

A (named list of) character strings giving effect size estimates.

statistic

A (named list of) character strings giving test statistic, parameters, and p values.

full_report

A (named list of) character strings comprised of estimate and statistic for each factor.

table

A data.frame containing all results; can, for example, be passed to apa_table.

[Package papaja version 0.1.0.9997 Index]