arrange_anova {papaja}	R Documentation

Create variance table from various ANOVA objects

Description

These methods take objects from various R functions that calculate ANOVA to create a data.frame containing a variance table. This function is not exported.

Usage

arrange_anova(x, ...)

## S3 method for class 'anova'
arrange_anova(x)

## S3 method for class 'summary.aov'
arrange_anova(x)

## S3 method for class 'summary.Anova.mlm'
arrange_anova(x, correction = "GG")

Arguments

x
Output object. See details.

...
Further arguments to pass to methods.

correction
Character. For summary.Anova.mlm objects, specifies the type of sphericity correction to be used. Either GG for Greenhouse-Geisser or HF for Huyn-Feldt methods or none is also possible. Ignored for other objects.

Details

The returned data.frame can be passed to functions such as print_anova.

Currently, methods for the following objects are available:

summary.aov

summary.aovlist

Anova.mlm

Value

data.frame of class apa_variance_table or apa_model_comp.

See Also

print_anova, print_model_comp

Examples

 ## Not run: 
   ## From Venables and Ripley (2002) p. 165.
   npk_aov <- aov(yield ~ block + N * P * K, npk)
   arrange_anova(summary(npk_aov))
 
## End(Not run)

[Package papaja version 0.1.0.9997 Index]