arrange_regression {papaja}	R Documentation

Create regression table

Description

These methods take glm or lm objects to create a data.frame containing a regression table. This function is not exported.

Usage

arrange_regression(x, est_name, standardized, ci, ...)

Arguments

x
glm or lm object. See details.

est_name
Character. If NULL (default) the name given in x (or a formally correct adaptation, such as "b^*" instead of "b" for standardized regression coefficients) is used, otherwise the supplied name is used. See details.

standardized
Logical. Indicates if coefficients are standardized or unstandardized and leading zeros are omitted if appropriate. See details.

ci
Numeric. Either a single value (range [0, 1]) giving the confidence level or a two-column matrix with confidence region bounds as column names (e.g. "2.5 %" and "97.5 %") and coefficient names as row names (in the same order as they appear in summary(x)$coefficients. See details.

Value

data.frame of class apa_regression_table.

[Package papaja version 0.1.0.9997 Index]