brighten {papaja}	R Documentation

Brighten up a colour

Description

We use this internal utility function to brighten up a specified colour (e.g., the swarm colour in apa_beeplot).

Usage

brighten(col, factor)

Arguments

col
Colour to be brightened up. Can be anything that is digestible by col2rgb.

factor
Numeric The rate with which the color channel should be boosted. Must be in the range between 0 and 1.

Value

A character vector as returned by rgb.

[Package papaja version 0.1.0.9997 Index]