cite_r {papaja}	R Documentation

Cite R and R-packages

Description

Creates a character string to cite R and R-packages.

Usage

cite_r(
  file = NULL,
  prefix = "R-",
  footnote = FALSE,
  pkgs = NULL,
  withhold = TRUE
)

Arguments

file
Character. The path and name of the .bib-file holding the references. If NULL only R is cited.

prefix
Character. The prefix used for all R-package reference handles.

footnote
Logical. Indicates if packages should be cited in a footnote. Ignored if no package information is available.

pkgs
Character. Vector of package names to cite or withhold depending on withhold.

withhold
Logical. If TRUE, pkgs constitutes a list of packages not to cite (a blacklist). If FALSE, pkgs constitutes a list of packages to cite (a whitelist).

Details

If footnote = FALSE a character string citing R and R-packages including version numbers is returned. Otherwise a named list with the elements r and pkgs is returned. The former element holds a character string citing R and a reference to a footnote; the latter element contains a character string for the footnote citing R-packages. For correct rendering, the footnote string needs to be a separate paragraph in the Markdown document.

Value

If footnote = FALSE a character string is returned, otherwise a named list with the elements r and pkgs.

See Also

r_refs, write_bib

Examples

cite_r()

[Package papaja version 0.1.0.9997 Index]