conf_int {papaja}	R Documentation

Between-subjects confidence intervals

Description

Calculates the deviation that is needed to construct confidence intervals for a vector of observations.

Usage

conf_int(x, level = 0.95, na.rm = TRUE)

conf.int(x, level = 0.95, na.rm = TRUE)

ci(x, level = 0.95, na.rm = TRUE)

Arguments

x
Numeric. A vector of observations from your dependent variable.

level
Numeric. Defines the width of the interval if confidence intervals are plotted. Defaults to 0.95 for 95% confidence intervals.

na.rm
Logical. Specifies if missing values are removed.

[Package papaja version 0.1.0.9997 Index]