fast_aggregate {papaja}	R Documentation

Aggregate data much faster using dplyr

Description

This is a convenience wrapper for aggregating your data using dplyr functions that tend to be much faster than the usual aggregate command. It is also easy to call from within a function. This function is not exported.

Usage

fast_aggregate(data, factors, dv, fun, na.rm = TRUE)

Arguments

data
A data.frame that contains the data.

factors
Character. A vector of factor names to aggregate data by.

dv
Character. The dependent variable to aggregate. All variables in data that contain this substring will be aggregated separately.

fun
Closure. The function used for aggregation.

Examples

NULL

[Package papaja version 0.1.0.9997 Index]