fetch_web_refs {papaja}	R Documentation

Fetch a .bib-reference file from the web (defunct)

Description

Downloads and saves a .bib-reference file form the web, so it can be used to cite references in a Markdown-document using pandoc or LaTeX. This function has been defunct. Please use download instead.

Usage

fetch_web_refs(x, bib_name)

Arguments

x
Character. URL of the .bib-file to fetch.

bib_name
Character. The path and name of the file to be created.

Details

If the function is called in an RMarkdown-document the file name specified as bib_name can be used in the YAML header as bibliography

See Also

cite_r, r_refs, write_bib

Examples

NULL

[Package papaja version 0.1.0.9997 Index]