format_cells {papaja}	R Documentation

Format numeric table cells

Description

Format all numeric cells of a table using printnum. This function is not exported.

Usage

format_cells(x, format.args = NULL)

Arguments

x
data.frame or matrix.

format.args
List. A named list of arguments to be passed to printnum to format numeric values.

See Also

printnum

Examples

NULL

[Package papaja version 0.1.0.9997 Index]