in_paren {papaja}	R Documentation

Replace parentheses with brackets

Description

Takes a single character or a list of characters and replaces parentheses with brackets. Can be used to prepare a string of statistics (e.g. containing degrees of freedom) for reporting within paretheses.

Usage

in_paren(x)

Arguments

x
Character. Single character or list of characters.

Value

An object of the same type as x.

See Also

apa_print

Examples

t_stat <- t.test(extra ~ group, data = sleep)
t_test_res <- apa_print(t_stat)
in_paren(t_test_res$stat)

[Package papaja version 0.1.0.9997 Index]