indent_stubs {papaja}	R Documentation

Add stub indentation

Description

Indents stubs by line and adds section headings This function is not exported.

Usage

indent_stubs(x, lines, filler = "   ")

Arguments

x
data.frame.

lines
List. A named list of vectors of length 2 giving the first and second row to indent. Names of list elements will be used as titles for indented sections.

filler
Character. Symbols used to indent stubs.

See Also

apa_table

Examples

NULL

[Package papaja version 0.1.0.9997 Index]