merge_tables {papaja}	R Documentation

Merge tables in list

Description

Takes a list of containing one or more matrix or data.frame and merges them into a single table. This function is not exported and currently unused.

Usage

merge_tables(x, empty_cells, row_names, added_stub_head)

Arguments

x
List. A named list containing one or more matrix or data.frame.

empty_cells
Character. String to place in empty cells; should be "" if the target document is LaTeX and "&nbsp;" if the target document is Word.

row_names
Logical. Vector of boolean values specifying whether to print column names for the corresponding list element.

added_stub_head
Character. Vector of names for first unnamed columns. See apa_table.

See Also

apa_table

Examples

NULL

[Package papaja version 0.1.0.9997 Index]