papaja {papaja}	R Documentation

APA manuscripts from R Markdown

Description

papaja is a R-package in the making including a R Markdown template that can be used with (or without) RStudio to produce documents, which conform to the American Psychological Association (APA, 2010) manuscript guidelines (6th Edition). The package uses the LaTeX document class apa6 and a .docx-reference file, so you can create PDF documents, or Word documents if you have to. Moreover, 'papaja' supplies R-functions (e.g., apa_print) that facilitate reporting results of your analyses in accordance with APA guidelines.

Details

Note, at this point 'papaja' is in active development and should be considered alpha. If you experience any problems, please open an issue with a reproducible example of the encountered problem.

Functions

To get a list of all functions try ls("package:papaja") or click the Index link at the bottom of this document.

System requirements

To use papaja you need to make sure the following software is installed on your computer:

RStudio (>= 0.98.932) or if you don't use RStudio , you need to install pandoc using the instructions for your operating system

A TeX distribution (2013 or later; e.g., MikTeX for Windows, MacTeX for Mac, obviously, or TeX Live for Linux)

Windows users should use MikTeX if possible. Currently, pandoc and the Windows version of Tex Live don't seem to like each other. Make sure you install the complete—not the basic—version.

Ubuntu 14.04 users need a few additional TeX packages for the document class apa6 to work:

sudo apt-get install texlive texlive-publishers texlive-fonts-extra texlive-latex-extra texlive-humanities lmodern

Authors

Frederik Aust (frederik.aust at uni-koeln.de). Marius Barth (marius.barth at uni-koeln.de).

Maintainer

Frederik Aust (frederik.aust at uni-koeln.de).

Contributors

Birk Diedenhofen

Christoph Stahl

References

American Psychological Association. (2010). Publication Manual of the American Psychological Association (6th edition). Washington, DC: American Psychological Association.

[Package papaja version 0.1.0.9997 Index]