print_interval {papaja}	R Documentation

Create interval estimate string

Description

Creates a character string to report an interval estimate. This function is not exported.

Usage

print_interval(x, conf_level = NULL, use_math = FALSE, interval_type, ...)

Arguments

x
Numeric. Either a vector of length 2 with attribute conf.level or a two-column matrix and confidence region bounds as column names (e.g. "2.5 %" and "97.5 %") and coefficient names as row names.

conf_level
Numeric. Vector of length 2 giving the lower and upper bounds of the confidence region in case they cannot be determined from column names or attributes of x.

use_math
Logical. Indicates whether to insert $ into the output so that Inf or scientific

interval_type
Character. Abbreviation indicating the type of interval estimate, e.g. CI.

...
Arguments passed on to printnum

See Also

printnum

Examples

## Not run: 
print_confint(c(1, 2), conf_level = 0.95)

## End(Not run)

[Package papaja version 0.1.0.9997 Index]