printp {papaja}	R Documentation

Prepare numeric values for printing as p value

Description

Convenience wrapper for printnum.numeric to print p values.

Usage

printp(x, digits = 3L, na_string = "", add_equals = FALSE)

Arguments

x
Numeric. Can be either a single value, vector, or matrix.

digits
Integer. The desired number of digits after the decimal point, passed on to formatC.

na_string
Character. String to print if element of x is NA.

add_equals
Logical. Indicates if the output string should be prepended with an equals sign.

Examples

printp(0.05)
printp(0.0005)
printp(0.99999999)

[Package papaja version 0.1.0.9997 Index]