render_appendix {papaja}	R Documentation

Render appendix

Description

This functions renders an R Markdown document without YAML header to a TeX fragment inside an appendix environment.

Usage

render_appendix(
  x,
  bibliography = rmarkdown::metadata$bibliography,
  csl = rmarkdown::metadata$csl,
  quiet = TRUE,
  ...
)

Arguments

x
Character. Input file name.

bibliography
Character. Location of bibliography file(s) to use.

csl
Character. Location of CSL file to use. Defaults to APA-style.

quiet
Logical. Suppresses pandoc command line output; see render. If FALSE output will be included in the document.

...
Arguments passed on to rmarkdown::pandoc_convert

input

Character vector containing paths to input files (files must be UTF-8 encoded)

to

Format to convert to (if not specified, you must specify output)

from

Format to convert from (if not specified then the format is determined based on the file extension of input).

output

Output file (if not specified then determined based on format being converted to).

citeproc

TRUE to run the pandoc-citeproc filter (for processing citations) as part of the conversion.

options

Character vector of command line options to pass to pandoc.

verbose

TRUE to show the pandoc command line which was executed

wd

Working directory in which code will be executed. If not supplied, defaults to the common base directory of input.

Details

By default x is converted to a TeX file which can be included in an R Markdown document as include:

   output:
     pdf_document:
       include:
         after_body: appendix.tex
   

If render_appendix is called form an R Markdown document with a target document type other than a PDF file a Markdown document is created and a warning returned (includes are only supported in PDF documents).

Default chunk options and hooks are set to those used in the R Markdown document from which render_appendix is called; otherwise defaults of md_document are used. It is, therefore, recommended to include render_appendix in your parent document.

[Package papaja version 0.1.0.9997 Index]