sel {papaja}	R Documentation

Select parameters

Description

If a list holds vectors of parameter values, this function extracts the i-th parameter value from each vector and creates a new list with these values. Especially helpful if a function is call repeatedly via do.call with different parameter values from within a function.

Usage

sel(x, i)

Arguments

x
List. A list of parameter values

i
Integer. The i-th element of each vector that is to be extracted.

Examples

NULL

[Package papaja version 0.1.0.9997 Index]