theme_apa {papaja}	R Documentation

APA compatible ggplot2 theme

Description

A ggplot2 theme with a white panel background, no grid lines, large axis and legend titles, and increased text padding for better readability.

Usage

theme_apa(base_size = 12, base_family = "", box = FALSE)

Arguments

base_size
Numeric. Base font size; other font sizes and margins are adjusted relative to this.

base_family
Character. Base font family.

box
Logical. Indicates whether to draw a box around the plot.

Details

This theme is an adaptation of theme_bw. In ggplot2, themes set the general aspect of the plot such as the color of the background, gridlines, the size and colour of fonts.

Examples

 ## Not run: 
   # Copied from ?ggtheme
   p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour = factor(gear))) + facet_wrap(~ am)
   p
   p + theme_apa()
 
## End(Not run)

[Package papaja version 0.1.0.9997 Index]