variable_label {papaja}	R Documentation

Assign or Extract Variable Labels

Description

Assign or extract variable labels of a vector or the columns (i.e., vectors) of a data.frame.

Usage

variable_label(x, ...)

## Default S3 method:
variable_label(x, ...)

## S3 method for class 'data.frame'
variable_label(x, ...)

variable_label(x) <- value

## Default S3 replacement method:
variable_label(x) <- value

## S3 replacement method for class 'data.frame'
variable_label(x) <- value

label_variable(x, ...)

variable_labels(x, ...)

variable_labels(x) <- value

label_variables(x, ...)

Arguments

x
Either a vector or a data.frame.

...
Further arguments that can be passes to methods.

value
Character. The variable label(s) to be assigned. If variable_label is applied to a single vector, this should be a length-one argument. If applied to a data.frame, value is required to be a named vector. Check the examples for details.

[Package papaja version 0.1.0.9997 Index]