word_title_page {papaja}	R Documentation

Create title page

Description

Creates text for the title and abstract page for MS Word documents. This function is not exported.

Usage

word_title_page(x)

Arguments

x
List. Meta data of the document as a result from yaml.load.

See Also

apa6_word

[Package papaja version 0.1.0.9997 Index]